(#58) move as many tests to unittests as possible
